Remove ConnectNetwotkHelper and merge with other network classes
