+++ app/src/org/commcare/connect/network/NetworkUtils.kt
@@ -1,8 +1,19 @@
 package org.commcare.connect.network
 
+import com.google.common.collect.ArrayListMultimap
+import com.google.gson.Gson
+import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.RequestBody
+import okhttp3.RequestBody.Companion.toRequestBody
+import org.commcare.connect.network.base.BaseApiHandler
+import org.commcare.core.network.ModernHttpRequester
 import org.commcare.util.LogTypes
+import org.commcare.utils.GlobalErrorUtil
+import org.commcare.utils.GlobalErrors
+import org.commcare.utils.optStringSafe
 import org.javarosa.core.io.StreamsUtil
 import org.javarosa.core.services.Logger
+import org.json.JSONException
 import org.json.JSONObject
 import retrofit2.HttpException
 import retrofit2.Response
@@ -89,4 +100,93 @@ object NetworkUtils {
         )
         Logger.exception(errorMessage, t)
     }
+
+
+    @JvmStatic
+    fun addVersionHeader(
+        headers: HashMap<String, String>,
+        version: String,
+    ) {
+        headers["Accept"] = "application/json;version=$version"
+    }
+
+    @JvmStatic
+    fun buildPostFormHeaders(
+        params: HashMap<String, Any>,
+        useFormEncoding: Boolean,
+        version: String,
+        outputHeaders: HashMap<String, String>,
+    ): RequestBody {
+        val requestBody: RequestBody
+
+        if (useFormEncoding) {
+            val multimap = ArrayListMultimap.create<String, String>()
+            for ((key, value) in params) {
+                multimap.put(key, value.toString())
+            }
+            requestBody = ModernHttpRequester.getPostBody(multimap)
+            outputHeaders.putAll(getContentHeadersForXFormPost(requestBody))
+        } else {
+            val json = Gson().toJson(params)
+            requestBody = json.toRequestBody("application/json".toMediaType())
+        }
+
+        addVersionHeader(outputHeaders, version)
+        return requestBody
+    }
+
+    private fun getContentHeadersForXFormPost(postBody: RequestBody): HashMap<String, String> {
+        val headers = HashMap<String, String>()
+        headers["Content-Type"] = "application/x-www-form-urlencoded"
+        try {
+            headers["Content-Length"] = postBody.contentLength().toString()
+        } catch (_: Exception) {
+        }
+        return headers
+    }
+
+    @JvmStatic
+    fun checkForLoginFromDifferentDevice(errorBody: String?): Boolean {
+        if (errorBody == null) return false
+        return try {
+            val json = JSONObject(errorBody)
+            "LOGIN_FROM_DIFFERENT_DEVICE" == json.optStringSafe("error_code", null)
+        } catch (_: JSONException) {
+            false
+        }
+    }
+
+    @JvmStatic
+    fun mapHttpErrorCode(
+        responseCode: Int,
+        errorBody: String?,
+    ): BaseApiHandler.PersonalIdOrConnectApiErrorCodes =
+        when (responseCode) {
+            401 -> {
+                BaseApiHandler.PersonalIdOrConnectApiErrorCodes.FAILED_AUTH_ERROR
+            }
+
+            403 -> {
+                BaseApiHandler.PersonalIdOrConnectApiErrorCodes.FORBIDDEN_ERROR
+            }
+
+            429 -> {
+                BaseApiHandler.PersonalIdOrConnectApiErrorCodes.RATE_LIMIT_EXCEEDED_ERROR
+            }
+
+            400 -> {
+                if (checkForLoginFromDifferentDevice(errorBody)) {
+                    GlobalErrorUtil.triggerGlobalError(GlobalErrors.PERSONALID_LOGIN_FROM_DIFFERENT_DEVICE_ERROR)
+                }
+                BaseApiHandler.PersonalIdOrConnectApiErrorCodes.BAD_REQUEST_ERROR
+            }
+
+            in 500..509 -> {
+                BaseApiHandler.PersonalIdOrConnectApiErrorCodes.SERVER_ERROR
+            }
+
+            else -> {
+                BaseApiHandler.PersonalIdOrConnectApiErrorCodes.UNKNOWN_ERROR
+            }
+        }
 }
